Bluej

+++ intObj/tp02/ej05/Rueda.java
@@ -4,35 +4,35 @@ public class Rueda {
     private float velocidad;
     private int radio;
 
-	public boolean isGirando() {
-		return girando;
-	}
-	public void setGirando(boolean girando) {
-		this.girando = girando;
-	}
+    public boolean isGirando() {
+        return girando;
+    }
+    public void setGirando(boolean girando) {
+        this.girando = girando;
+    }
 
-	public int getRpm() {
-		return rpm;
-	}
-	public void setRpm(int rpm) {
-		this.rpm = rpm;
-	}
+    public int getRpm() {
+        return rpm;
+    }
+    public void setRpm(int rpm) {
+        this.rpm = rpm;
+    }
 
-	public float getVelocidad() {
-		return velocidad;
-	}
-	public void setVelocidad(float velocidad) {
-		this.velocidad = velocidad;
-	}
+    public float getVelocidad() {
+        return velocidad;
+    }
+    public void setVelocidad(float velocidad) {
+        this.velocidad = velocidad;
+    }
 
-	public int getRadio() {
-		return radio;
-	}
-	public void setRadio(int radio) {
-		this.radio = radio;
-	}
+    public int getRadio() {
+        return radio;
+    }
+    public void setRadio(int radio) {
+        this.radio = radio;
+    }
 
-    public float calcVelocidad() {
+    public double calcVelocidad() {
         if(this.isGirando()) {
             return ((this.getRpm() * 2*Math.PI)/60) * this.getRadio();
         } else {
